requirements.txt
Django
django-cors-headers
djangorestframework
whitenoise
mssql-django
pyodbc
python-dotenv
django-nonrelated-inlines